Do the durability bar, damage, and ammo sums in long as well
Three more places did int arithmetic on durability-scale values: the
durability bar percentage (dur * 100 wraps once a huge tool has worn
more than ~21M points), the damage total in AbilityHelper.damageTool (a
death penalty of TotalDurability / 10 on such a tool can wrap the sum
negative, which the zero-guard then reads as a full repair), and
AmmoItem.addAmmo (a restock from an endgame material can saturate its
float and wrap the count). The Flux energy bar in the same method had
the identical shape and is widened alongside it.

+++ src/main/java/tconstruct/library/weaponry/AmmoItem.java
@@ -51,7 +51,7 @@ public abstract class AmmoItem extends ToolCore implements IBattlegearWeapon, IA
         if (!stack.hasTagCompound()) return toAdd;
         NBTTagCompound tags = stack.getTagCompound().getCompoundTag("InfiTool");
         int oldCount = tags.getInteger("Ammo");
-        int newCount = Math.min(oldCount + toAdd, getMaxAmmo(stack));
+        int newCount = (int) Math.min((long) oldCount + toAdd, getMaxAmmo(stack));
         setAmmo(newCount, stack);
         return toAdd - (newCount - oldCount);
     }
